requirements.txt
langchain==0.1.17
langchain_core==0.1.50
langchain_openai==0.1.6
langchain_text_splitters==0.0.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.32.2
